booking-api/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Google Calendar API
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-auth==2.25.2

# Google Cloud
google-cloud-secret-manager==2.17.0

# 設定ファイル
PyYAML==6.0.1

# 日時処理
python-dateutil==2.8.2

# バリデーション
email-validator==2.1.0

# ログ
python-json-logger==2.0.7
